+++ packages/article-magazine-standard/src/article-magazine-standard.web.js
@@ -68,7 +68,8 @@ class ArticlePage extends Component {
       spotAccountId,
       paidContentClassName,
       isPreview,
-      additionalRelatedArticlesFlag
+      additionalRelatedArticlesFlag,
+      algoliaSearchKeys
     } = this.props;
 
     if (error || isLoading) {
@@ -87,6 +88,7 @@ class ArticlePage extends Component {
         paidContentClassName={paidContentClassName}
         isPreview={isPreview}
         additionalRelatedArticlesFlag={additionalRelatedArticlesFlag}
+        algoliaSearchKeys={algoliaSearchKeys}
       />
     );
   }