+++ packages/article-byline/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.5.30](https://github.com/newsuk/times-components/compare/@times-components/article-byline@3.5.29...@times-components/article-byline@3.5.30) (2021-04-28)
+
+**Note:** Version bump only for package @times-components/article-byline
+
+
+
+
+
 ## [3.5.29](https://github.com/newsuk/times-components/compare/@times-components/article-byline@3.5.28...@times-components/article-byline@3.5.29) (2021-04-22)
 
 **Note:** Version bump only for package @times-components/article-byline