+++ packages/article-paragraph/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.153](https://github.com/newsuk/times-components/compare/@times-components/article-paragraph@1.4.152...@times-components/article-paragraph@1.4.153) (2021-04-28)
+
+**Note:** Version bump only for package @times-components/article-paragraph
+
+
+
+
+
 ## [1.4.152](https://github.com/newsuk/times-components/compare/@times-components/article-paragraph@1.4.151...@times-components/article-paragraph@1.4.152) (2021-04-22)
 
 **Note:** Version bump only for package @times-components/article-paragraph