+++ packages/article-topics/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.2.8](https://github.com/newsuk/times-components/compare/@times-components/article-topics@4.2.7...@times-components/article-topics@4.2.8) (2021-04-28)
+
+**Note:** Version bump only for package @times-components/article-topics
+
+
+
+
+
 ## [4.2.7](https://github.com/newsuk/times-components/compare/@times-components/article-topics@4.2.6...@times-components/article-topics@4.2.7) (2021-04-22)
 
 **Note:** Version bump only for package @times-components/article-topics