+++ packages/article/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [7.7.0](https://github.com/newsuk/times-components/compare/@times-components/article@7.6.2...@times-components/article@7.7.0) (2021-04-28)
+
+
+### Features
+
+* **TE-64:** algolia related articles ([#2686](https://github.com/newsuk/times-components/issues/2686)) ([f3bc502](https://github.com/newsuk/times-components/commit/f3bc5028b0059f2ce08d02bdfce4acbea3b012bf))
+
+
+
+
+
 ## [7.6.2](https://github.com/newsuk/times-components/compare/@times-components/article@7.6.1...@times-components/article@7.6.2) (2021-04-22)
 
 **Note:** Version bump only for package @times-components/article