+++ packages/date-publication/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.21.55](https://github.com/newsuk/times-components/compare/@times-components/date-publication@0.21.54...@times-components/date-publication@0.21.55) (2021-04-28)
+
+**Note:** Version bump only for package @times-components/date-publication
+
+
+
+
+
 ## [0.21.54](https://github.com/newsuk/times-components/compare/@times-components/date-publication@0.21.53...@times-components/date-publication@0.21.54) (2021-04-22)
 
 **Note:** Version bump only for package @times-components/date-publication