+++ packages/error-view/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.54](https://github.com/newsuk/times-components/compare/@times-components/error-view@2.3.53...@times-components/error-view@2.3.54) (2021-04-28)
+
+**Note:** Version bump only for package @times-components/error-view
+
+
+
+
+
 ## [2.3.53](https://github.com/newsuk/times-components/compare/@times-components/error-view@2.3.52...@times-components/error-view@2.3.53) (2021-04-22)
 
 **Note:** Version bump only for package @times-components/error-view