+++ packages/tracking/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.7.7](https://github.com/newsuk/times-components/compare/@times-components/tracking@2.7.6...@times-components/tracking@2.7.7) (2021-04-28)
+
+**Note:** Version bump only for package @times-components/tracking
+
+
+
+
+
 ## [2.7.6](https://github.com/newsuk/times-components/compare/@times-components/tracking@2.7.5...@times-components/tracking@2.7.6) (2021-04-22)
 
 **Note:** Version bump only for package @times-components/tracking