+++ packages/video/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.11.3](https://github.com/newsuk/times-components/compare/@times-components/video@4.11.2...@times-components/video@4.11.3) (2021-04-28)
+
+**Note:** Version bump only for package @times-components/video
+
+
+
+
+
 ## [4.11.2](https://github.com/newsuk/times-components/compare/@times-components/video@4.11.1...@times-components/video@4.11.2) (2021-04-22)
 
 **Note:** Version bump only for package @times-components/video